Fix build errors - project now builds successfully
Fixed Issues:
- Ambiguous cos/sin functions in AnimatedLogoView - added CGFloat cast
- Task naming conflicts - changed to _Concurrency.Task throughout
- onChange API compatibility - removed oldValue parameter for iOS 16 support
- All files now compile without errors

Build Status: ✅ BUILD SUCCEEDED on iPhone 17 Pro simulator

+++ ios-swift/Notelayer/Notelayer/Views/Shared/AnimatedLogoView.swift
@@ -72,8 +72,8 @@ struct AnimatedLogoView: View {
             let particle = ConfettiParticle(
                 id: UUID(),
                 offset: CGSize(
-                    width: cos(radians) * distance,
-                    height: sin(radians) * distance
+                    width: CGFloat(cos(radians)) * distance,
+                    height: CGFloat(sin(radians)) * distance
                 ),
                 color: colors.randomElement() ?? .blue,
                 size: CGFloat.random(in: 4...8),
